requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
httpx>=0.23.0
pydantic>=1.10.0
python-dotenv>=0.21.0
